+++ src/libromdata/Handheld/Nintendo3DSFirm.cpp
@@ -228,7 +228,7 @@ int Nintendo3DSFirm::loadFieldData(void)
 
 	// Nintendo 3DS firmware binary header.
 	const N3DS_FIRM_Header_t *const firmHeader = &d->firmHeader;
-	d->fields->reserve(5);	// Maximum of 5 fields.
+	d->fields->reserve(6);	// Maximum of 6 fields.
 
 	// Read the firmware binary.
 	unique_ptr<uint8_t[]> firmBuf;
@@ -320,24 +320,31 @@ int Nintendo3DSFirm::loadFieldData(void)
 			unsigned int searchlen;	// Search string length, without the NULL terminator.
 		};
 		static const Arm9VerStr_t arm9VerStr[] = {
-			{"Luma3DS", "Luma3DS v", 9},
-			{"GodMode9", "GodMode9 Explorer v", 19},
-			{"Decrypt9WIP", "Decrypt9WIP (", 13},
-			{"Hourglass9", "Hourglass9 v", 12},
+			{"Luma3DS",		"Luma3DS v", 9},
+			{"GodMode9",		"GodMode9 Explorer v", 19},	// Older versions
+			{"GodMode9",		"GodMode9 v", 10},		// Newer versions (v1.9.1; TODO check for first one?)
+			{"Decrypt9WIP",		"Decrypt9WIP (", 13},
+			{"Hourglass9",		"Hourglass9 v", 12},
+			{"ntrboot_flasher",	"ntrboot_flasher: %s", 19},	// version info isn't hard-coded
+			{"SafeB9SInstaller",	"SafeB9SInstaller v", 18},
+			{"OpenFirmInstaller",	"OpenFirmInstaller v", 19},
+			{"fastboot3DS",		"fastboot3DS v", 13},
+
+			{nullptr, nullptr, 0}
 		};
 
 		const char *arm9VerStr_title = nullptr;
 		string s_verstr;
-		for (unsigned int i = 0; i < ARRAY_SIZE(arm9VerStr); i++) {
+		for (const Arm9VerStr_t *p = arm9VerStr; p->title != nullptr; p++) {
 			const char *verstr = static_cast<const char*>(memmem(
-				firmBuf.get(), szFile, arm9VerStr[i].searchstr, arm9VerStr[i].searchlen));
+				firmBuf.get(), szFile, p->searchstr, p->searchlen));
 			if (!verstr)
 				continue;
 
-			arm9VerStr_title = arm9VerStr[i].title;
+			arm9VerStr_title = p->title;
 
 			// Version does NOT include the 'v' character.
-			verstr += arm9VerStr[i].searchlen;
+			verstr += p->searchlen;
 			const char *end = (const char*)firmBuf.get() + szFile;
 			int count = 0;
 			while (verstr < end && count < 32 && verstr[count] != 0 &&
@@ -357,6 +364,44 @@ int Nintendo3DSFirm::loadFieldData(void)
 			break;
 		}
 
+		// Sighax status.
+		// TODO: If it's SPI, we need to decrypt the FIRM contents.
+		// Reference: https://github.com/TuxSH/firmtool/blob/master/firmtool/__main__.py
+		const uint32_t first4 = be32_to_cpu(*(reinterpret_cast<const uint32_t*>(firmHeader->signature)));
+		struct SighaxStatus_t {
+			uint32_t first4;
+			const char *status;
+		};
+		static const SighaxStatus_t sighaxStatus[] = {
+			{0xB6724531,	"NAND retail"},		// SciresM
+			{0x6EFF209C,	"NAND retail"},		// sighax.com
+			{0x88697CDC,	"NAND devkit"},		// SciresM
+
+			{0x6CF52F89,	"NCSD retail"},
+			{0x53CB0E4E,	"NCSD devkit"},
+
+			{0x37E96B10,	"SPI retail"},
+			{0x18722BC7,	"SPI devkit"},
+
+			{0, nullptr}
+		};
+
+		const char *s_sighax_status = nullptr;
+		for (const SighaxStatus_t *p = sighaxStatus; p->first4 != 0; p++) {
+			if (p->first4 == first4) {
+				s_sighax_status = p->status;
+				break;
+			}
+		}
+
+		if (s_sighax_status) {
+			// Sighaxed. Assume it's ARM9 homebrew.
+			firmBinDesc = C_("Nintendo3DSFirm", "ARM9 Homebrew");
+		} else {
+			// Not sighaxed.
+			s_sighax_status = C_("Nintendo3DSFirm", "Not sighaxed");
+		}
+
 		// Add the firmware type field.
 		d->fields->addField_string(C_("Nintendo3DSFirm", "Type"),
 			(firmBinDesc ? firmBinDesc : C_("RomData", "Unknown")));
@@ -369,6 +414,8 @@ int Nintendo3DSFirm::loadFieldData(void)
 		if (!s_verstr.empty()) {
 			d->fields->addField_string(C_("RomData", "Version"), s_verstr);
 		}
+
+		d->fields->addField_string(C_("Nintendo3DSFirm", "Sighax Status"), s_sighax_status);
 	} else {
 		// Add the firmware type field.
 		d->fields->addField_string(C_("Nintendo3DSFirm", "Type"),